+++ include/nfx/serialization/json/SerializableDocument.h
@@ -29,10 +29,10 @@
 
 #pragma once
 
-#include <nfx/json/Document.h>
-
+#include "traits/SerializationTraits.h"
 #include "Concepts.h"
-#include "SerializationTraits.h"
+
+#include <nfx/json/Document.h>
 
 namespace nfx::serialization::json
 {